requirements.txt
aniso8601==9.0.1
APScheduler==3.7.0
attrs==20.3.0
autopep8==1.5.6
beautifulsoup4==4.9.3
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cymem==2.0.5
decorator==4.4.2
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
Flask-MySQL==1.5.2
flask-restplus==0.13.0
idna==2.10
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
lxml==4.6.3
MarkupSafe==1.1.1
murmurhash==1.0.5
networkx==2.5
newspaper3k==0.2.8
nltk==3.5
numpy==1.20.1
packaging==20.9
pathy==0.4.0
Pillow==8.1.2
plac==1.1.3
preshed==3.0.5
pycodestyle==2.7.0
pydantic==1.7.3
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-crfsuite==0.9.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
pyvi==0.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
schedule==1.0.0
scikit-learn==0.24.1
scipy==1.6.1
sgmllib3k==1.0.0
six==1.15.0
sklearn-crfsuite==0.3.6
smart-open==3.0.0
soupsieve==2.2.1
srsly==1.0.5
tabulate==0.8.9
thinc==7.4.5
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
toolz==0.11.1
tqdm==4.59.0
typer==0.3.2
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
voluptuous==0.12.1
w3lib==1.22.0
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==1.0.1
